+++ src/typings/types.ts
@@ -6,6 +6,7 @@ export type EndPoint = `/${string}`;
 export type NumberResolvable = number | `${number}`;
 export type Page<T> = T[];
 export type BaseFetchPageOptions = Partial<Pick<PageFetchOptions, 'perPage' | 'page'>>;
+export type ListingAction = 'bump' | 'publish' | 'unpublish';
 export type ListingCategory = 'cfp' | 'forhire' | 'collabs' | 'education' | 'jobs' | 'mentors' | 'products' | 'mentees' | 'forsale' | 'events' | 'misc';
 export type Tags = string;
 export type TypeOfArticle = 'article' | 'video_article';